Added CORS configuration to allowe front-end UI requests

+++ src/main/java/com/haleha/repository/AccountsRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AccountsRepository  extends CrudRepository<Accounts, Integer> {
+public interface AccountsRepository  extends CrudRepository<Accounts, Long> {
     Accounts findByCustomerId(int customerId);
 }
